server/converter/requirements.txt
Pillow>=11.0.0
imageio>=2.30.0
pillow-heif>=0.15.0
yt-dlp>=2025.7.0
